Nothing

+++ Restaurant/Rest/Cook.cpp
@@ -104,7 +104,7 @@ void Cook::serveOrder(Order* _order, int& timeStep) {
 	_order->setStatus(SRV);
 	_order->Set_serveTime(timeStep);
 	// we now calculate the time taken to serve the order:
-	float timeTaken = ceil(_order->get_size() / speed); // the size of the order divided by the speed of the cook
+	float timeTaken = ceil(float(_order->get_size() / speed)); // the size of the order divided by the speed of the cook
 														   // then rounded to the nearest time step
 	cooldownEnd = timeTaken + timeStep;
 	_order->Set_finishTime(_order->Get_servetime() + timeTaken);
